requirements.txt
alabaster==0.7.12
attrs==19.3.0
awscli==1.16.272
Babel==2.7.0
backcall==0.1.0
botocore==1.13.8
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==4.5.4
cycler==0.10.0
decorator==4.4.2
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
idna==2.8
imagesize==1.1.0
importlib-metadata==1.5.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.2.0
numpy==1.17.3
packaging==20.3
pandas==0.25.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.5.2
pluggy==0.13.1
prompt-toolkit==3.0.3
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.1
pyaml==20.3.1
pyasn1==0.4.7
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
pyparsing==2.4.5
pytest==5.4.1
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
retrying==1.3.3
rsa==3.4.2
s3transfer==0.2.1
scikit-learn==0.21.3
scikit-optimize==0.7.4
scipy==1.3.1
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.6
wcwidth==0.1.8
zipp==2.2.0
